# Speed up (filtered) KNN queries for flat vector fields