--- a/server/src/main/java/org/elasticsearch/index/mapper/vectors/DenseVectorFieldMapper.java
+++ b/server/src/main/java/org/elasticsearch/index/mapper/vectors/DenseVectorFieldMapper.java
@@ -30,9 +30,13 @@ import org.apache.lucene.index.SegmentReadState;
 import org.apache.lucene.index.SegmentWriteState;
 import org.apache.lucene.index.VectorEncoding;
 import org.apache.lucene.index.VectorSimilarityFunction;
+import org.apache.lucene.search.BooleanClause;
+import org.apache.lucene.search.BooleanQuery;
 import org.apache.lucene.search.FieldExistsQuery;
 import org.apache.lucene.search.Query;
 import org.apache.lucene.search.join.BitSetProducer;
+import org.apache.lucene.search.join.ScoreMode;
+import org.apache.lucene.search.join.ToParentBlockJoinQuery;
 import org.apache.lucene.search.knn.KnnSearchStrategy;
 import org.apache.lucene.util.BitUtil;
 import org.apache.lucene.util.BytesRef;
@@ -1391,6 +1395,18 @@ public class DenseVectorFieldMapper extends FieldMapper {
         public final int hashCode() {
             return Objects.hash(type, doHashCode());
         }
+
+        /**
+         * Indicates whether the underlying vector search is performed using a flat (exhaustive) approach.
+         * <p>
+         * When {@code true}, it means the search does not use any approximate nearest neighbor (ANN)
+         * acceleration structures such as HNSW or IVF. Instead, it performs a brute-force comparison
+         * against all candidate vectors. This information can be used by higher-level components
+         * to decide whether additional acceleration or optimization is necessary.
+         *
+         * @return {@code true} if the vector search is flat (exhaustive), {@code false} if it uses ANN structures
+         */
+        abstract boolean isFlat();
     }
 
     abstract static class QuantizedIndexOptions extends DenseVectorIndexOptions {
@@ -1763,6 +1779,11 @@ public class DenseVectorFieldMapper extends FieldMapper {
         }
 
         @Override
+        boolean isFlat() {
+            return true;
+        }
+
+        @Override
         public boolean updatableTo(DenseVectorIndexOptions update) {
             return update.type.equals(this.type)
                 || update.type.equals(VectorIndexType.HNSW)
@@ -1809,6 +1830,11 @@ public class DenseVectorFieldMapper extends FieldMapper {
         @Override
         public int doHashCode() {
             return Objects.hash(type);
+        }
+
+        @Override
+        boolean isFlat() {
+            return true;
         }
     }
 
@@ -1858,6 +1884,11 @@ public class DenseVectorFieldMapper extends FieldMapper {
         @Override
         public int doHashCode() {
             return Objects.hash(m, efConstruction, confidenceInterval, rescoreVector);
+        }
+
+        @Override
+        boolean isFlat() {
+            return false;
         }
 
         @Override
@@ -1932,6 +1963,11 @@ public class DenseVectorFieldMapper extends FieldMapper {
         }
 
         @Override
+        boolean isFlat() {
+            return true;
+        }
+
+        @Override
         public String toString() {
             return "{type=" + type + ", confidence_interval=" + confidenceInterval + ", rescore_vector=" + rescoreVector + "}";
         }
@@ -1997,6 +2033,11 @@ public class DenseVectorFieldMapper extends FieldMapper {
         @Override
         public int doHashCode() {
             return Objects.hash(m, efConstruction, confidenceInterval, rescoreVector);
+        }
+
+        @Override
+        boolean isFlat() {
+            return false;
         }
 
         @Override
@@ -2089,6 +2130,11 @@ public class DenseVectorFieldMapper extends FieldMapper {
         }
 
         @Override
+        boolean isFlat() {
+            return false;
+        }
+
+        @Override
         public String toString() {
             return "{type=" + type + ", m=" + m + ", ef_construction=" + efConstruction + "}";
         }
@@ -2124,6 +2170,11 @@ public class DenseVectorFieldMapper extends FieldMapper {
         @Override
         int doHashCode() {
             return Objects.hash(m, efConstruction, rescoreVector);
+        }
+
+        @Override
+        boolean isFlat() {
+            return false;
         }
 
         @Override
@@ -2177,6 +2228,11 @@ public class DenseVectorFieldMapper extends FieldMapper {
         @Override
         int doHashCode() {
             return CLASS_NAME_HASH;
+        }
+
+        @Override
+        boolean isFlat() {
+            return true;
         }
 
         @Override
@@ -2235,6 +2291,11 @@ public class DenseVectorFieldMapper extends FieldMapper {
         @Override
         int doHashCode() {
             return Objects.hash(clusterSize, defaultNProbe, rescoreVector);
+        }
+
+        @Override
+        boolean isFlat() {
+            return false;
         }
 
         @Override
@@ -2485,9 +2546,21 @@ public class DenseVectorFieldMapper extends FieldMapper {
             KnnSearchStrategy searchStrategy
         ) {
             elementType.checkDimensions(dims, queryVector.length);
-            Query knnQuery = parentFilter != null
-                ? new ESDiversifyingChildrenByteKnnVectorQuery(name(), queryVector, filter, k, numCands, parentFilter, searchStrategy)
-                : new ESKnnByteVectorQuery(name(), queryVector, k, numCands, filter, searchStrategy);
+            Query knnQuery;
+            if (indexOptions.isFlat()) {
+                knnQuery = filter == null
+                    ? createExactKnnBitQuery(queryVector)
+                    : new BooleanQuery.Builder().add(createExactKnnBitQuery(queryVector), BooleanClause.Occur.SHOULD)
+                        .add(filter, BooleanClause.Occur.FILTER)
+                        .build();
+                if (parentFilter != null) {
+                    knnQuery = new ToParentBlockJoinQuery(knnQuery, parentFilter, ScoreMode.Max);
+                }
+            } else {
+                knnQuery = parentFilter != null
+                    ? new ESDiversifyingChildrenByteKnnVectorQuery(name(), queryVector, filter, k, numCands, parentFilter, searchStrategy)
+                    : new ESKnnByteVectorQuery(name(), queryVector, k, numCands, filter, searchStrategy);
+            }
             if (similarityThreshold != null) {
                 knnQuery = new VectorSimilarityQuery(
                     knnQuery,
@@ -2513,9 +2586,22 @@ public class DenseVectorFieldMapper extends FieldMapper {
                 float squaredMagnitude = VectorUtil.dotProduct(queryVector, queryVector);
                 elementType.checkVectorMagnitude(similarity, ElementType.errorByteElementsAppender(queryVector), squaredMagnitude);
             }
-            Query knnQuery = parentFilter != null
-                ? new ESDiversifyingChildrenByteKnnVectorQuery(name(), queryVector, filter, k, numCands, parentFilter, searchStrategy)
-                : new ESKnnByteVectorQuery(name(), queryVector, k, numCands, filter, searchStrategy);
+
+            Query knnQuery;
+            if (indexOptions.isFlat()) {
+                knnQuery = filter == null
+                    ? createExactKnnByteQuery(queryVector)
+                    : new BooleanQuery.Builder().add(createExactKnnByteQuery(queryVector), BooleanClause.Occur.SHOULD)
+                        .add(filter, BooleanClause.Occur.FILTER)
+                        .build();
+                if (parentFilter != null) {
+                    knnQuery = new ToParentBlockJoinQuery(knnQuery, parentFilter, ScoreMode.Max);
+                }
+            } else {
+                knnQuery = parentFilter != null
+                    ? new ESDiversifyingChildrenByteKnnVectorQuery(name(), queryVector, filter, k, numCands, parentFilter, searchStrategy)
+                    : new ESKnnByteVectorQuery(name(), queryVector, k, numCands, filter, searchStrategy);
+            }
             if (similarityThreshold != null) {
                 knnQuery = new VectorSimilarityQuery(
                     knnQuery,
@@ -2568,7 +2654,16 @@ public class DenseVectorFieldMapper extends FieldMapper {
                 numCands = Math.max(adjustedK, numCands);
             }
             Query knnQuery;
-            if (indexOptions instanceof BBQIVFIndexOptions bbqIndexOptions) {
+            if (indexOptions.isFlat()) {
+                knnQuery = filter == null
+                    ? createExactKnnFloatQuery(queryVector)
+                    : new BooleanQuery.Builder().add(createExactKnnFloatQuery(queryVector), BooleanClause.Occur.SHOULD)
+                        .add(filter, BooleanClause.Occur.FILTER)
+                        .build();
+                if (parentFilter != null) {
+                    knnQuery = new ToParentBlockJoinQuery(knnQuery, parentFilter, ScoreMode.Max);
+                }
+            } else if (indexOptions instanceof BBQIVFIndexOptions bbqIndexOptions) {
                 knnQuery = parentFilter != null
                     ? new DiversifyingChildrenIVFKnnFloatVectorQuery(
                         name(),
@@ -2594,11 +2689,12 @@ public class DenseVectorFieldMapper extends FieldMapper {
                     : new ESKnnFloatVectorQuery(name(), queryVector, adjustedK, numCands, filter, knnSearchStrategy);
             }
             if (rescore) {
-                knnQuery = new RescoreKnnVectorQuery(
+                knnQuery = RescoreKnnVectorQuery.fromInnerQuery(
                     name(),
                     queryVector,
                     similarity.vectorSimilarityFunction(indexVersionCreated, ElementType.FLOAT),
                     k,
+                    adjustedK,
                     knnQuery
                 );
             }
@@ -2624,7 +2720,7 @@ public class DenseVectorFieldMapper extends FieldMapper {
             return elementType;
         }
 
-        public IndexOptions getIndexOptions() {
+        public DenseVectorIndexOptions getIndexOptions() {
             return indexOptions;
         }