--- a/server/src/main/java/org/elasticsearch/search/vectors/DenseVectorQuery.java
+++ b/server/src/main/java/org/elasticsearch/search/vectors/DenseVectorQuery.java
@@ -207,4 +207,5 @@ public abstract class DenseVectorQuery extends Query {
             return iterator.docID();
         }
     }
+
 }