… docs/changelog/130251.yaml
… server/src/test/java/org/elasticsearch/index/mapper/DynamicMappingTests.java
… server/src/test/java/org/elasticsearch/index/mapper/vectors/DenseVectorFieldTypeTests.java
… server/src/test/java/org/elasticsearch/search/vectors/AbstractKnnVectorQueryBuilderTestCase.java
… server/src/test/java/org/elasticsearch/search/vectors/DiversifyingParentBlockQueryTests.java