@jimczi 2025-06-30
I tweaked the brute-force nested version so that it always diversifies the child docs. It’s similar in spirit to what we do for HNSW, but much faster here since we can just walk through each parent’s block in order.

I ran a comparison between `main` and this branch using the `nested` mode in the `random_vector` track (added [[here](https://github.com/elastic/rally-tracks/pull/804)](https://github.com/elastic/rally-tracks/pull/804)):

```
|                                                Min Throughput | brute-force-filtered-search | 2052.39        |  85.0901      | -1967.3    |  ops/s |   -95.85% |
|                                               Mean Throughput | brute-force-filtered-search | 2214.2         |  87.3678      | -2126.83   |  ops/s |   -96.05% |
|                                             Median Throughput | brute-force-filtered-search | 2240.51        |  87.3315      | -2153.18   |  ops/s |   -96.10% |
|                                                Max Throughput | brute-force-filtered-search | 2250.5         |  88.3819      | -2162.12   |  ops/s |   -96.07% |
|                                       50th percentile latency | brute-force-filtered-search |    2.2724      |  79.5413      |    77.2689 |     ms | +3400.33% |
|                                       90th percentile latency | brute-force-filtered-search |    2.49625     | 136.674       |   134.178  |     ms | +5375.18% |
|                                       99th percentile latency | brute-force-filtered-search |    8.14283     | 199.337       |   191.195  |     ms | +2348.01% |
|                                     99.9th percentile latency | brute-force-filtered-search |   50.8515      | 265.505       |   214.654  |     ms |  +422.12% |
|                                    99.99th percentile latency | brute-force-filtered-search |   70.5917      | 359.416       |   288.824  |     ms |  +409.15% |
|                                      100th percentile latency | brute-force-filtered-search |   74.4209      | 438.119       |   363.698  |     ms |  +488.70% |
|                                  50th percentile service time | brute-force-filtered-search |    2.2724      |  79.5413      |    77.2689 |     ms | +3400.33% |
|                                  90th percentile service time | brute-force-filtered-search |    2.49625     | 136.674       |   134.178  |     ms | +5375.18% |
|                                  99th percentile service time | brute-force-filtered-search |    8.14283     | 199.337       |   191.195  |     ms | +2348.01% |
|                                99.9th percentile service time | brute-force-filtered-search |   50.8515      | 265.505       |   214.654  |     ms |  +422.12% |
|                               99.99th percentile service time | brute-force-filtered-search |   70.5917      | 359.416       |   288.824  |     ms |  +409.15% |
|                                 100th percentile service time | brute-force-filtered-search |   74.4209      | 438.119       |   363.698  |     ms |  +488.70% |
|                                                    error rate | brute-force-filtered-search |    0           |   0           |     0      |      % |     0.00% |
```

As you can see, the performance drops off pretty hard in the nested filtered case, lots of room for improvement. I think we should tackle that in a follow-up. From profiling, stuff like `TimeOutCheckingBits` is showing up a lot, so there’s some easy wins we can go after.