@jimczi 2025-06-30
> OK, if this passes given the new yaml tests, I think this is g2g!

Well it's broken ;) 
Let me check what's going on.