@jimczi 2025-06-30
Ok that was an issue when the diversifying query hits [NO_MORE_DOCS](https://github.com/elastic/elasticsearch/pull/130251/commits/62215c426cfa2463abcd01b673c256e616bd8c5e).
Let's wait for this another round.